sentinelweb-backup/backend/requirements.txt
fastapi==0.115.12  # Updated to latest version
uvicorn==0.34.0  # Updated from 0.22.0 for security
pydantic==2.10.6  # Updated from 1.10.7 for security
sqlalchemy==2.0.38  # Updated from 2.0.12 for security
asyncpg==0.29.0  # Updated from 0.27.0 for security
httpx==0.27.0  # Updated from 0.24.0 for security
# Replacing python-jose with PyJWT due to vulnerabilities CVE-2024-33664 and CVE-2024-33663
# # python-jose==3.4.0  # Removed due to vulnerabilities CVE-2024-33664 and CVE-2024-33663
pyjwt==2.8.0  # Using a secure version that doesn't have CVE-2024-53861  # Downgraded to fix CVE-2024-53861
passlib==1.7.4
python-multipart==0.0.20  # Updated to latest version to fix CVE-2024-539810  # Updated to latest version to fix CVE-2024-53981
redis==5.0.3  # Updated from 4.5.4 for security
websockets==12.0  # Updated from 11.0.3 for security
python-dotenv==1.0.1  # Updated from 1.0.0 for security
bcrypt==4.3.0  # Updated from 4.0.1 for security
aiofiles==23.2.0  # Updated from 23.1.0 for security
pykml==0.2.0
gpxpy==1.5.0

# Security
cryptography==46.0.0  # Updated to latest version to fix CVE-2024-26130, CVE-2024-12797, CVE-2024-6119  # Updated from 44.0.1 to fix CVE-2024-26130, CVE-2024-12797, CVE-2024-6119
pyopenssl==24.0.0  # Secure version
